Introduce a secondary grouping key for rollout operation
The way rollout-operator currently is designed in percludes deploying
all of the LGTM stack to a single namespace (which is required in some
circumstances). This adds an optional additional label that can be used
to employ a secondary grouping to allow for all components to be
deployed in a single namespace and have the rollout-operator execute as
expected. For example:

Loki ingester:
```
apiVersion: apps/v1
kind: StatefulSet
metadata:
  labels:
    name: loki-ingester-zone-a
    rollout-group: ingester
    rollout-secondary-group: loki
```

Mimir ingester:
```
apiVersion: apps/v1
kind: StatefulSet
metadata:
  labels:
    name: mimir-ingester-zone-a
    rollout-group: ingester
    rollout-secondary-group: mimir
```